extract part of apple gradle actions to separated classes
