+++ resources-generator/src/main/kotlin/dev/icerock/gradle/generator/apple/AppleMRGenerator.kt
@@ -12,12 +12,14 @@ import com.squareup.kotlinpoet.STRING
 import com.squareup.kotlinpoet.TypeSpec
 import dev.icerock.gradle.MultiplatformResourcesPluginExtension
 import dev.icerock.gradle.generator.MRGenerator
+import dev.icerock.gradle.generator.apple.action.CopyResourcesFromKLibsToExecutableAction
+import dev.icerock.gradle.generator.apple.action.CopyResourcesFromKLibsToFrameworkAction
+import dev.icerock.gradle.generator.apple.action.PackResourcesToKLibAction
 import dev.icerock.gradle.tasks.CopyExecutableResourcesToApp
 import dev.icerock.gradle.tasks.CopyFrameworkResourcesToAppEntryPointTask
 import dev.icerock.gradle.tasks.CopyFrameworkResourcesToAppTask
 import dev.icerock.gradle.utils.calculateResourcesHash
 import dev.icerock.gradle.utils.dependsOnProcessResources
-import dev.icerock.gradle.utils.klibs
 import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.Task
@@ -37,14 +39,7 @@ import org.jetbrains.kotlin.gradle.tasks.FatFrameworkTask
 import org.jetbrains.kotlin.gradle.tasks.FrameworkDescriptor
 import org.jetbrains.kotlin.gradle.tasks.KotlinNativeCompile
 import org.jetbrains.kotlin.gradle.tasks.KotlinNativeLink
-import org.jetbrains.kotlin.konan.file.zipDirAs
-import org.jetbrains.kotlin.library.impl.KotlinLibraryLayoutImpl
 import java.io.File
-import java.io.InputStream
-import java.util.Properties
-import java.util.zip.ZipEntry
-import java.util.zip.ZipException
-import java.util.zip.ZipFile
 import kotlin.reflect.full.memberProperties
 
 @Suppress("TooManyFunctions")
@@ -114,65 +109,14 @@ class AppleMRGenerator(
         val compileTask: KotlinNativeCompile = compilation.compileKotlinTask
         compileTask.dependsOn(generationTask)
 
-        // lambda will broke gradle UP-TO-DATE mark!
-        @Suppress("ObjectLiteralToLambda")
-        compileTask.doLast(object : Action<Task> {
-            override fun execute(task: Task) {
-                task as KotlinNativeCompile
-
-                val klibFile = task.outputFile.get()
-                val repackDir = File(klibFile.parent, klibFile.nameWithoutExtension)
-                val defaultDir = File(repackDir, "default")
-                val resRepackDir = File(defaultDir, "resources")
-
-                unzipTo(zipFile = klibFile, outputDirectory = repackDir)
-
-                val manifestFile = File(defaultDir, "manifest")
-                val manifest = Properties()
-                manifest.load(manifestFile.inputStream())
-
-                val uniqueName = manifest["unique_name"] as String
-
-                val loadableBundle = LoadableBundle(
-                    directory = resRepackDir,
-                    bundleName = uniqueName,
-                    developmentRegion = baseLocalizationRegion,
-                    identifier = bundleIdentifier
-                )
-                loadableBundle.write()
-
-                assetsDirectory?.let { assetsDir ->
-                    val process = Runtime.getRuntime().exec(
-                        "xcrun actool Assets.xcassets --compile . --platform iphoneos --minimum-deployment-target 9.0",
-                        emptyArray(),
-                        assetsDir.parentFile
-                    )
-                    val errors = process.errorStream.bufferedReader().readText()
-                    val input = process.inputStream.bufferedReader().readText()
-                    val result = process.waitFor()
-                    if (result != 0) {
-                        println("can't compile assets - $result")
-                        println(input)
-                        println(errors)
-                    } else {
-                        assetsDir.deleteRecursively()
-                    }
-                }
-
-                resourcesGenerationDir.copyRecursively(
-                    loadableBundle.resourcesDir,
-                    overwrite = true
-                )
-
-                val repackKonan = org.jetbrains.kotlin.konan.file.File(repackDir.path)
-                val klibKonan = org.jetbrains.kotlin.konan.file.File(klibFile.path)
-
-                klibFile.delete()
-                repackKonan.zipDirAs(klibKonan)
-
-                repackDir.deleteRecursively()
-            }
-        })
+        compileTask.doLast(
+            PackResourcesToKLibAction(
+                baseLocalizationRegion = baseLocalizationRegion,
+                bundleIdentifier = bundleIdentifier,
+                assetsDirectory = assetsDirectory,
+                resourcesGenerationDir = resourcesGenerationDir,
+            )
+        )
     }
 
     private fun setupFrameworkResources() {
@@ -186,14 +130,7 @@ class AppleMRGenerator(
 
                 val linkTask = framework.linkTask
 
-                // lambda will broke gradle UP-TO-DATE mark!
-                @Suppress("ObjectLiteralToLambda")
-                linkTask.doLast(object : Action<Task> {
-                    override fun execute(task: Task) {
-                        task as KotlinNativeLink
-                        copyKlibsResourcesIntoFramework(task)
-                    }
-                })
+                linkTask.doLast(CopyResourcesFromKLibsToFrameworkAction())
 
                 if (framework.isStatic) {
                     val resourcesExtension =
@@ -206,44 +143,6 @@ $linkTask produces static framework, Xcode should have Build Phase with copyFram
                         )
                     }
                     createCopyFrameworkResourcesTask(linkTask)
-                }
-            }
-    }
-
-    private fun copyKlibsResourcesIntoFramework(linkTask: KotlinNativeLink) {
-        val project = linkTask.project
-        val framework = linkTask.binary as Framework
-
-        copyResourcesFromLibraries(
-            linkTask = linkTask,
-            project = project,
-            outputDir = framework.outputFile
-        )
-    }
-
-    private fun copyResourcesFromLibraries(
-        linkTask: KotlinNativeLink,
-        project: Project,
-        outputDir: File
-    ) {
-        linkTask.klibs
-            .filter { it.extension == "klib" }
-            .filter { it.exists() }
-            .forEach { inputFile ->
-                project.logger.info("copy resources from $inputFile into $outputDir")
-                val klibKonan = org.jetbrains.kotlin.konan.file.File(inputFile.path)
-                val klib = KotlinLibraryLayoutImpl(klib = klibKonan, component = "default")
-                val layout = klib.extractingToTemp
-
-                try {
-                    File(layout.resourcesDir.path).copyRecursively(
-                        target = outputDir,
-                        overwrite = true
-                    )
-                } catch (@Suppress("SwallowedException") exc: NoSuchFileException) {
-                    project.logger.info("resources in $inputFile not found")
-                } catch (@Suppress("SwallowedException") exc: java.nio.file.NoSuchFileException) {
-                    project.logger.info("resources in $inputFile not found (empty lib)")
                 }
             }
     }
@@ -292,27 +191,14 @@ $linkTask produces static framework, Xcode should have Build Phase with copyFram
 
     private fun setupTestsResources() {
         val kotlinNativeTarget = compilation.target as KotlinNativeTarget
-        val project = kotlinNativeTarget.project
 
         kotlinNativeTarget.binaries
             .matching { it is TestExecutable && it.compilation.associateWith.contains(compilation) }
             .configureEach {
                 val executable = it as TestExecutable
-
                 val linkTask = executable.linkTask
 
-                // lambda will broke gradle UP-TO-DATE mark!
-                @Suppress("ObjectLiteralToLambda")
-                linkTask.doLast(object : Action<Task> {
-                    override fun execute(task: Task) {
-                        task as KotlinNativeLink
-                        copyResourcesFromLibraries(
-                            linkTask = task,
-                            project = project,
-                            outputDir = executable.outputDirectory
-                        )
-                    }
-                })
+                linkTask.doLast(CopyResourcesFromKLibsToExecutableAction())
             }
     }
 
@@ -362,36 +248,6 @@ $linkTask produces static framework, Xcode should have Build Phase with copyFram
         project.tasks.withType(FatFrameworkTask::class)
             .configureEach { it.doLast(fatAction) }
     }
-
-    private fun unzipTo(outputDirectory: File, zipFile: File) {
-        ZipFile(zipFile).use { zip ->
-            val outputDirectoryCanonicalPath = outputDirectory.canonicalPath
-            for (entry in zip.entries()) {
-                unzipEntryTo(outputDirectory, outputDirectoryCanonicalPath, zip, entry)
-            }
-        }
-    }
-
-    private fun unzipEntryTo(
-        outputDirectory: File,
-        outputDirectoryCanonicalPath: String,
-        zip: ZipFile,
-        entry: ZipEntry
-    ) {
-        val output = outputDirectory.resolve(entry.name)
-        if (!output.canonicalPath.startsWith(outputDirectoryCanonicalPath)) {
-            throw ZipException("Zip entry '${entry.name}' is outside of the output directory")
-        }
-        if (entry.isDirectory) {
-            output.mkdirs()
-        } else {
-            output.parentFile.mkdirs()
-            zip.getInputStream(entry).use { it.copyTo(output) }
-        }
-    }
-
-    private fun InputStream.copyTo(file: File): Long =
-        file.outputStream().use { copyTo(it) }
 
     companion object {
         const val BUNDLE_PROPERTY_NAME = "bundle"
