+++ resources-generator/src/main/kotlin/dev/icerock/gradle/utils/ZipUtils.kt
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2023 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package dev.icerock.gradle.utils
+
+import java.io.File
+import java.io.InputStream
+import java.util.zip.ZipEntry
+import java.util.zip.ZipException
+import java.util.zip.ZipFile
+
+internal fun unzipTo(outputDirectory: File, zipFile: File) {
+    ZipFile(zipFile).use { zip ->
+        val outputDirectoryCanonicalPath = outputDirectory.canonicalPath
+        for (entry in zip.entries()) {
+            unzipEntryTo(outputDirectory, outputDirectoryCanonicalPath, zip, entry)
+        }
+    }
+}
+
+private fun unzipEntryTo(
+    outputDirectory: File,
+    outputDirectoryCanonicalPath: String,
+    zip: ZipFile,
+    entry: ZipEntry
+) {
+    val output = outputDirectory.resolve(entry.name)
+    if (!output.canonicalPath.startsWith(outputDirectoryCanonicalPath)) {
+        throw ZipException("Zip entry '${entry.name}' is outside of the output directory")
+    }
+    if (entry.isDirectory) {
+        output.mkdirs()
+    } else {
+        output.parentFile.mkdirs()
+        zip.getInputStream(entry).use { it.copyTo(output) }
+    }
+}
+
+private fun InputStream.copyTo(file: File): Long =
+    file.outputStream().use { copyTo(it) }
